requirements.txt
aiohttp[speedups] >= 3.6
click
errata-tool >= 1.22
future
koji >= 1.18
kobo ~= 0.19.0
rpm-py-installer
semver
pygit2 == 1.6.*
python-bugzilla >= 3.2
pyyaml
requests
requests_kerberos >= 0.13
ruamel.yaml
six
typing ; python_version <= '3.4'
tenacity
